Rename a few files to fit the naming scheme
