+++ web/client/src/components/infrastructure/highlight-helpers.ts
