+++ web/client/src/components/navigation/station-search/station-search.vue
@@ -86,7 +86,7 @@
 
 <script setup lang="ts">
 import SoroButton from '@/components/soro-button.vue';
-import { ElementTypeLabels } from '@/components/infrastructure/elementTypes';
+import { ElementTypeLabels } from '@/components/infrastructure/element-types';
 import { iconUrl, iconExtension } from '@/components/infrastructure/add-icons';
 </script>
 
@@ -94,7 +94,7 @@ import { iconUrl, iconExtension } from '@/components/infrastructure/add-icons';
 import { defineComponent } from 'vue';
 import { InfrastructureNamespace } from '@/stores/infrastructure-store';
 import { mapActions, mapMutations, mapState } from 'vuex';
-import { ElementType } from '@/components/infrastructure/elementTypes';
+import { ElementType } from '@/components/infrastructure/element-types';
 
 const validSearchTypes = [
     ElementType.STATION,
